requirements.txt
pillow
exif
imagehash
numpy